% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{assemble_chunk}
\alias{assemble_chunk}
\title{Assemble a Chunk}
\usage{
assemble_chunk(chunk)
}
\arguments{
\item{chunk}{a chunk in output from `parse_dat`}
}
\value{
a multidimensional array
}
\description{
Combine multiple sections of a chunk to get a complete representation of it's blocks
}
