% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk_top_block}
\alias{chunk_top_block}
\title{Top Block in a Chunk}
\usage{
chunk_top_block(chunk)
}
\arguments{
\item{chunk}{a chunk in output from `parse_dat`}
}
\value{
a 16 x 16 matrix
}
\description{
Get the highest opaque block in a chunk, after assembling it.
}
\details{
This is the block we want to see when plotting.
}
