% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganvil.R
\docType{data}
\name{color_map}
\alias{color_map}
\title{Block Color Map}
\format{
A data frame with 1395 rows and 3 variables:
\describe{
  \item{block_name}{Internal minecraft blockname. Links to chunk palette.}
  \item{base_text}{Base texture that the block maps to}
  \item{rgb}{Mapped or hand picked rgb value for block}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
color_map
}
\description{
A data.frame for mapping minecraft block names to RGB colors. This is helpful
for creating map plots that line up with in game colors.
}
\details{
The RGB values are pretty much improvised. There's json files in
assets/minecraft/models/block/ (in the minecraft.jar file) that seem to map
block names to the base textures in assets/minecraft/textures/block/, which
are all pngs. For blocks with available png files, the average RGB value is
used. Otherwise, the RGB value may be hand picked or missing.
}
\keyword{datasets}
