% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_region_name}
\alias{get_region_name}
\title{Get Region Name}
\usage{
get_region_name(chunk_x, chunk_z)
}
\arguments{
\item{chunk_x}{The chunk X coordinate (east-west)}

\item{chunk_z}{The chunk z coordinate (north-south)}
}
\value{
A character of length one containing the filename
}
\description{
Get the name of a region file (.mca file) based on the chunk x and z coordinates.
These coordinates can be determined by using the F3 menu in game.
The files are found under word/region/ for a given world.
}
