% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{parse_dat}
\alias{parse_dat}
\title{Parse a region file}
\usage{
parse_dat(fn)
}
\arguments{
\item{fn}{File path of .dat file}
}
\value{
a list of header data and NBT data
}
\description{
Based on documentation from [here](https://wiki.vg/Region_Files)
}
