% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBT.R
\name{parse_nbt}
\alias{parse_nbt}
\title{Parse uncompressed NBT data}
\usage{
parse_nbt(connection, N = 1)
}
\arguments{
\item{connection}{file path of uncompressed NBT file}

\item{N}{number of tags to read}
}
\value{
a nested list of parsed results
}
\description{
Most Minecraft data is stored in NBT format. This is a binary format similar to json
The format is documented [here](https://wiki.vg/NBT)
All NBT files contain all data in a base `TAG_Compund` tag,
so the default value of `N = 1` will typically read all tags.
}
\details{
Be careful with `TAG_Long`. R doesn't support 64 bit integers. So instead,
the bytes are just read as a vector of 8 unsigned integers.
}
