% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBT.R
\name{parse_nbt_file}
\alias{parse_nbt_file}
\title{Parse a raw NBT file}
\usage{
parse_nbt_file(file_name, N = 1)
}
\arguments{
\item{file_name}{file path}

\item{N}{number of tags to read}
}
\value{
a nested list of parsed results
}
\description{
Parse a raw NBT file
}
