% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_top_blocks}
\alias{plot_top_blocks}
\title{Plot top blocks}
\usage{
plot_top_blocks(top_blocks)
}
\arguments{
\item{top_blocks}{A character matrix (2d array) of block names}
}
\value{
A ggplot2 plot
}
\description{
Plots a 2 dimensional character array of blocks. Designed for plotting
top blocks produced by `region_top_block`.
}
\details{
Notice the in game x is mapped to the x coordinate in ggplot so the
north-south axis will be displayed left to right.
}
