% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region_data}
\alias{region_data}
\title{Get Region Data}
\usage{
region_data(dat_file)
}
\arguments{
\item{dat_file}{File path of .dat file}
}
\value{
A data.frame with block data. It will be over a million rows.
}
\description{
Get every block in a region with x,y, and coordinates as a data frame.
Useful for doing analysis of the world file.
}
