% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region_top_block}
\alias{region_top_block}
\title{Top Blocks in a Region}
\usage{
region_top_block(dat_file)
}
\arguments{
\item{dat_file}{File path of .dat file}
}
\value{
A 2d character matrix
}
\description{
Get highest non-air block at each x-z coordinate in a region.
This results in a 512 x 512 character matrix useful for birds-eye plotting.
}
