% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{section_to_array}
\alias{section_to_array}
\title{Chunk Section to R Array}
\usage{
section_to_array(section)
}
\arguments{
\item{section}{A specific chunk section}
}
\value{
a 3d array
}
\description{
Convert chunk section data to a 3d array. Each element in the array
corresponds to a block from that section's palette. This format is
documented at [gamepedia](https://minecraft.gamepedia.com/Chunk_format)
}
